Initial implimentation of linear battery reports.
