+++ zhaquirks/centralite/3130.py
@@ -11,7 +11,8 @@ from zigpy.zcl.clusters.general import (
 )
 from zigpy.zcl.clusters.measurement import TemperatureMeasurement
 
-from zhaquirks.centralite import CENTRALITE, PowerConfigurationCluster
+from zhaquirks import PowerConfigurationCluster
+from zhaquirks.centralite import CENTRALITE
 from zhaquirks.const import (
     COMMAND,
     COMMAND_MOVE,
@@ -35,6 +36,17 @@ from zhaquirks.osram import OSRAM
 DIAGNOSTICS_CLUSTER_ID = 0x0B05  # decimal = 2821
 
 
+class CustomPowerConfigurationCluster(PowerConfigurationCluster):
+    """Custom PowerConfigurationCluster."""
+
+    cluster_id = PowerConfigurationCluster.cluster_id
+    MIN_VOLTS = 21
+    MAX_VOLTS = 31
+
+    def _update_attribute(self, attrid, value):
+        super()._update_attribute(attrid, value)
+
+
 class CentraLite3130(CustomDevice):
     """Custom device representing centralite 3130."""
 
@@ -50,7 +62,7 @@ class CentraLite3130(CustomDevice):
                 DEVICE_TYPE: zha.DeviceType.LEVEL_CONTROL_SWITCH,
                 INPUT_CLUSTERS: [
                     Basic.cluster_id,
-                    PowerConfigurationCluster.cluster_id,
+                    CustomPowerConfigurationCluster.cluster_id,
                     Identify.cluster_id,
                     PollControl.cluster_id,
                     TemperatureMeasurement.cluster_id,
@@ -70,7 +82,7 @@ class CentraLite3130(CustomDevice):
             1: {
                 INPUT_CLUSTERS: [
                     Basic.cluster_id,
-                    PowerConfigurationCluster,
+                    CustomPowerConfigurationCluster,
                     Identify.cluster_id,
                     PollControl.cluster_id,
                     DIAGNOSTICS_CLUSTER_ID,
