+++ zhaquirks/centralite/3300S.py
@@ -5,7 +5,8 @@ from zigpy.zcl.clusters.general import Basic, BinaryInput, Identify, Ota, PollCo
 from zigpy.zcl.clusters.measurement import TemperatureMeasurement
 from zigpy.zcl.clusters.security import IasZone
 
-from zhaquirks.centralite import CENTRALITE, PowerConfigurationCluster
+from zhaquirks import PowerConfigurationCluster
+from zhaquirks.centralite import CENTRALITE
 from zhaquirks.const import (
     DEVICE_TYPE,
     ENDPOINTS,
