+++ zhaquirks/centralite/3310S.py
@@ -9,7 +9,8 @@ from zigpy.quirks.smartthings import (
 from zigpy.zcl.clusters.general import Basic, Identify, Ota, PollControl
 from zigpy.zcl.clusters.measurement import TemperatureMeasurement
 
-from zhaquirks.centralite import CENTRALITE, PowerConfigurationCluster
+from zhaquirks import PowerConfigurationCluster
+from zhaquirks.centralite import CENTRALITE
 from zhaquirks.const import (
     DEVICE_TYPE,
     ENDPOINTS,
