+++ zhaquirks/centralite/3321S.py
@@ -6,10 +6,10 @@ from zigpy.zcl.clusters.general import Basic, Identify, Ota, PollControl
 from zigpy.zcl.clusters.measurement import TemperatureMeasurement
 from zigpy.zcl.clusters.security import IasZone
 
+from zhaquirks import PowerConfigurationCluster
 from zhaquirks.centralite import (
     CENTRALITE,
     CentraLiteAccelCluster,
-    PowerConfigurationCluster,
 )
 from zhaquirks.const import (
     DEVICE_TYPE,
