+++ zhaquirks/smartthings/moisturev4.py
@@ -5,7 +5,7 @@ from zigpy.zcl.clusters.general import Basic, BinaryInput, Identify, Ota, PollCo
 from zigpy.zcl.clusters.measurement import TemperatureMeasurement
 from zigpy.zcl.clusters.security import IasZone
 
-from zhaquirks.centralite import PowerConfigurationCluster
+from zhaquirks import PowerConfigurationCluster
 
 from . import SMART_THINGS
 from ..const import (
