+++ zhaquirks/smartthings/tag_v4.py
@@ -5,9 +5,8 @@ from zigpy.profiles import zha
 from zigpy.quirks import CustomDevice
 from zigpy.zcl.clusters.general import Basic, BinaryInput, Identify, Ota, PollControl
 
-from zhaquirks import Bus, LocalDataCluster
+from zhaquirks import Bus, LocalDataCluster, PowerConfigurationCluster
 
-from ..centralite import PowerConfigurationCluster
 from ..const import DEVICE_TYPE, ENDPOINTS, INPUT_CLUSTERS, OUTPUT_CLUSTERS, PROFILE_ID
 
 _LOGGER = logging.getLogger(__name__)
@@ -21,6 +20,8 @@ class FastPollingPowerConfigurationCluster(PowerConfigurationCluster):
     cluster_id = PowerConfigurationCluster.cluster_id
     FREQUENCY = 45
     MINIMUM_CHANGE = 1
+    # MIN_VOLTS = 15
+    # MAX_VOLTS = 28
 
     async def configure_reporting(
         self,
